requirements.txt
streamlit==1.28.1
langchain==0.0.350
langchain-google-genai==0.0.5
google-generativeai==0.3.2
pymysql==1.1.0
python-dotenv==1.0.0
pandas==2.1.3 